# resources/swagger.json
{"openapi":"3.0.0","paths":{"/v1/auth/signup":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthEntity"}}}}},"tags":["Auth"]}},"/v1/auth/signin":{"post":{"operationId":"AuthController_signin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthEntity"}}}}},"tags":["Auth"]}},"/v1/shops/{shopId}/reviews":{"get":{"operationId":"ReviewController_getShopReviews","parameters":[{"name":"shopId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReviewEntity"}}}}}},"tags":["Reviews"],"security":[{"bearer":[]}]},"post":{"operationId":"ReviewController_createReview","parameters":[{"name":"shopId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewEntity"}}}}},"tags":["Reviews"],"security":[{"bearer":[]}]}},"/v1/shops/{shopId}/reviews/page":{"get":{"operationId":"ReviewController_getShopReviewsByPage","parameters":[{"name":"shopId","required":true,"in":"path","schema":{"type":"string"}},{"name":"first","required":false,"in":"query","schema":{"type":"number"}},{"name":"last","required":false,"in":"query","schema":{"type":"number"}},{"name":"after","required":false,"in":"query","schema":{"type":"string"}},{"name":"before","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"PageResponseOfReviewEntity","allOf":[{"$ref":"#/components/schemas/Page"},{"properties":{"edges":{"type":"array","title":"EdgeOfReviewEntity","items":{"type":"object","required":["cursor","node"],"properties":{"cursor":{"type":"string"},"node":{"type":"object","$ref":"#/components/schemas/ReviewEntity"}}}}}}]}}}}},"tags":["Reviews"],"security":[{"bearer":[]}]}},"/v1/reviews":{"get":{"operationId":"ReviewController_getUserReviews","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReviewEntity"}}}}}},"tags":["Reviews"],"security":[{"bearer":[]}]}},"/v1/reviews/page":{"get":{"operationId":"ReviewController_getUserReviewsByPage","parameters":[{"name":"first","required":false,"in":"query","schema":{"type":"number"}},{"name":"last","required":false,"in":"query","schema":{"type":"number"}},{"name":"after","required":false,"in":"query","schema":{"type":"string"}},{"name":"before","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"PageResponseOfReviewEntity","allOf":[{"$ref":"#/components/schemas/Page"},{"properties":{"edges":{"type":"array","title":"EdgeOfReviewEntity","items":{"type":"object","required":["cursor","node"],"properties":{"cursor":{"type":"string"},"node":{"type":"object","$ref":"#/components/schemas/ReviewEntity"}}}}}}]}}}}},"tags":["Reviews"],"security":[{"bearer":[]}]}},"/v1/reviews/{reviewId}":{"get":{"operationId":"ReviewController_getReviewById","parameters":[{"name":"reviewId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewEntity"}}}}},"tags":["Reviews"],"security":[{"bearer":[]}]},"patch":{"operationId":"ReviewController_editReviewById","parameters":[{"name":"reviewId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewEntity"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewEntity"}}}}},"tags":["Reviews"],"security":[{"bearer":[]}]},"delete":{"operationId":"ReviewController_deleteReviewById","parameters":[{"name":"reviewId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Reviews"],"security":[{"bearer":[]}]}},"/v1/rooms":{"get":{"operationId":"RoomController_getAllRooms","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomEntity"}}}}}},"tags":["Rooms"],"security":[{"bearer":[]}]},"post":{"operationId":"RoomController_createRoom","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomEntity"}}}}}},"tags":["Rooms"],"security":[{"bearer":[]}]}},"/v1/rooms/user":{"get":{"operationId":"RoomController_getUserRooms","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomEntity"}}}}}},"tags":["Rooms"],"security":[{"bearer":[]}]}},"/v1/rooms/{roomId}":{"get":{"operationId":"RoomController_getRoomById","parameters":[{"name":"roomId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomEntity"}}}}}},"tags":["Rooms"],"security":[{"bearer":[]}]},"patch":{"operationId":"RoomController_editRoomById","parameters":[{"name":"roomId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomEntity"}}}},"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomEntity"}}}}}},"tags":["Rooms"],"security":[{"bearer":[]}]},"delete":{"operationId":"RoomController_deleteRoomById","parameters":[{"name":"roomId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Rooms"],"security":[{"bearer":[]}]}},"/v1/rooms/{roomId}/user/{userId}":{"post":{"operationId":"RoomController_addUserToRoom","parameters":[{"name":"roomId","required":true,"in":"path","schema":{"type":"string"}},{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Rooms"],"security":[{"bearer":[]}]},"delete":{"operationId":"RoomController_removeUserFromRoom","parameters":[{"name":"roomId","required":true,"in":"path","schema":{"type":"string"}},{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Rooms"],"security":[{"bearer":[]}]}},"/v1/shops":{"get":{"operationId":"ShopController_getShops","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ShopEntity"}}}}}},"tags":["Shops"],"security":[{"bearer":[]}]},"post":{"operationId":"ShopController_createShop","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShopDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShopEntity"}}}}},"tags":["Shops"],"security":[{"bearer":[]}]}},"/v1/shops/{shopId}":{"get":{"operationId":"ShopController_getShopById","parameters":[{"name":"shopId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShopEntity"}}}}},"tags":["Shops"],"security":[{"bearer":[]}]},"patch":{"operationId":"ShopController_editShopById","parameters":[{"name":"shopId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShopDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShopEntity"}}}}},"tags":["Shops"],"security":[{"bearer":[]}]},"delete":{"operationId":"ShopController_deleteShopById","parameters":[{"name":"shopId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Shops"],"security":[{"bearer":[]}]}},"/v1/users/me":{"get":{"operationId":"UserController_getMe","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/v1/users":{"patch":{"operationId":"UserController_editUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/v1/users/role":{"patch":{"operationId":"UserController_changeUserRole","parameters":[{"name":"role","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}}},"info":{"title":"NestJs Boilerplate API","description":"NestJs Boilerplate API","version":"1.0","contact":{}},"tags":[{"name":"NestJs","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AuthDto":{"type":"object","properties":{"email":{"type":"string","default":"email@email.com"},"password":{"type":"string"},"username":{"type":"string"},"name":{"type":"string"}},"required":["email","password","username","name"]},"AuthEntity":{"type":"object","properties":{"accessToken":{"type":"string"},"userId":{"type":"string"}},"required":["accessToken","userId"]},"ReviewEntity":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string","nullable":true},"description":{"type":"string","nullable":true},"rating":{"type":"number","nullable":true},"userId":{"type":"string"},"shopId":{"type":"string"}},"required":["id","title","description","rating","userId","shopId"]},"ReviewDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"rating":{"type":"number"}}},"UserEntity":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"},"name":{"type":"string","nullable":true}},"required":["id","username","email","role","name"]},"RoomEntity":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"description":{"type":"string","nullable":true},"link":{"type":"string"},"users":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}},"required":["id","title","description","link"]},"RoomDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"link":{"type":"string"}},"required":["title","link"]},"ShopEntity":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string","nullable":true},"website":{"type":"string","nullable":true}},"required":["id","name","description","website"]},"ShopDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"website":{"type":"string"}},"required":["name","description","website"]},"UserDto":{"type":"object","properties":{"email":{"type":"string","default":"email@email.com"},"username":{"type":"string"},"name":{"type":"string"}},"required":["email"]}}}}